Clearly identify ill-formed and invalid items, and specify processor behavior on them.
This heavily rewords the definition of additional data and the item's
"argument". It also moves some content from section 4 (Creating CBOR-Based
Protocols) into the main specification when it was actually describing when
data is invalid or ill-formed.

This does not clarify when tagged values are invalid: that's a separate PR.