Mention tag 257, which is more functional than tag 36.
Close #121

This is not meant to open the floodgates to all registrations that
were made since RFC 7049, but is intended to address one specific
defect of tag 36.

This isn't quite deprecating tag 36, but pointing out its limitations.

Instead of this change, we could repair tag 36 (e.g., allow both text
and byte strings), but that would mean less stability.